Finished Setting Up Music
Finished adding all of the music related services for now. All services including importing a song, saving the song file, saving the metatdata that i want to a json file that operates the same way as the old fake song list, the ability to delete songs. The only optimizations I could make is essentially reloading the page so that when you add or delete music files you dont have to got to the alarm page and then back to the music page to see your additions to your library as well as giving the user feedback through an alert dialog that you added or succeslfully deleted or imported music. Now I will be starting the alarm side of things which should be easier now that I basically have sort of an outline of how I want to do things since I did the music first.